requirements.txt
astroid==2.3.3
beautifulsoup4==4.8.2
dj-database-url==0.5.0
Django==2.2.10
django-appconf==1.0.3
django-bootstrap4==1.1.1
django-bootstrap-datepicker-plus==3.0.5
django-compat==1.0.15
django-constance==2.6.0
django-countries==5.5
django-dbbackup==3.2.0
django-filter==2.2.0
django-hijack==2.2.1
django-picklefield==2.1.1
django-post-office==3.4.0
django-select2==7.2.0
django-settings-export==1.2.1
djangorestframework==3.11.0
dodgy==0.2.1
isort==4.3.21
jsonfield==3.1.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysqlclient==1.4.6
pep8-naming==0.4.1
prospector==1.2.0
pycodestyle==2.4.0
pydocstyle==5.0.2
pyflakes==2.1.1
pylint==2.4.4
pylint-celery==0.3
pylint-django==2.0.12
pylint-flask==0.6
pylint-plugin-utils==0.6
python-decouple==3.3
pytz==2019.3
pyuca==1.2
PyYAML==5.3
requirements-detector==0.6
setoptconf==0.2.0
six==1.14.0
snowballstemmer==2.0.0
soupsieve==1.9.5
sqlparse==0.3.0
typed-ast==1.4.1
wrapt==1.11.2
